backend/requirements.txt
Django
django-admin-import
django-admin-sortable
django-admin-tools
django-ckeditor-widget
django-cors-headers
django-extensions
django-filter
django-jinja
django-jinja2
django-grappelli
django-imagekit
django-social-share
django-socialprofile
django-tagging
django-tinymce
django-ckeditor
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
djangorestframework-filters
django-rest-swagger
django-rest-registration
django-taggit 
django-WYMEditor
drf-url-filters
gunicorn
html2text
Jinja2
Markdown
psycopg2
Pillow
python-social-auth
pytz
pyyaml
social-auth-core
sqlparse
stripe
pytest
pytest-django
pytest-cov
social-auth-app-django
